Ignore it if we get a clearly invalid snitch_id.
This will make it easy to disable reporting to dms by ommitting snitch
ids on staging system. One can write code like the following.

    DeadMansSnitch.report(ENV['SNITCH_ID'])

With that one can set a snitch id on production and omit setting a
snitch id on staging to ensure we don't send requests from staging to
dms.